requirements.txt
streamlit
pandas
LogisticRegression
NearestNeighbors
matplotlib.pyplot
sklearn.linear_model
sklearn.neighbors